refactor: Rename files tftp_server.* to tftpSrv.* (u-t OK!)
Github issue #4
